+++ src/Common/StringTools.h
@@ -129,7 +129,7 @@ std::string base64Decode(std::string const& encoded_string);
 
 std::string ipAddressToString(uint32_t ip);
 bool parseIpAddress(uint32_t& ip, const std::string& addr);
-bool parseIpAddressAndPort(uint32_t& ip, uint32_t& port, const std::string& addr);
+bool parseIpAddressAndPort(uint32_t& ip, uint16_t &port, const std::string& addr);
 
 std::string timeIntervalToString(uint64_t intervalInSeconds);
