+++ src/P2p/ConnectionContext.h
@@ -36,7 +36,7 @@ struct CryptoNoteConnectionContext {
   uint8_t version;
   boost::uuids::uuid m_connection_id;
   uint32_t m_remote_ip = 0;
-  uint32_t m_remote_port = 0;
+  uint16_t m_remote_port = 0;
   bool m_is_income = false;
   time_t m_started = 0;
